Change reaction degeneracy algorithm
The structure of the algorithm is modified slightly, but the main
result is that identical reactions with different templates are
no longer marked as duplicates. This means that only 1 will make
it into the final model during a normal RMG job.

The reason for this is that different templates for identical
reactions should only result from different resonance structures.
Since resonance structures do not represent distinct states in
reality, these reactions should not be kept as duplicates.

The new implementation separates the reaction comparison and
processing steps. The isomorphic/identical/sameTemplate values are
determined first, and the appropriate action to take is determined
later. This allows different actions to be taken for reactions
that are identical/non-identical with different templates.